Remove "-rHEAD" from the try run options; use LKGR instead
HEAD seems to be fairly regularly broken these days, unfortunately.

The reason for having -rHEAD before was for when LKGR had not caught
up with a previous DEPS update (which would cause a patch to DEPS in
LKGR to fail to apply).

This shouldn't be a problem these days, though, since DEPS updates
don't tend to be frequent enough (i.e. multiple updates per day).
Even with one DEPS update per day, LKGR should be OK.